rename stream to match filename
